Postgis.Simple.Internal

Description

Low level operations on Postgis extention of PostgreSQL Database

User data classes

class GeoData a where

Type class for valid geometry data

geoType function must return valid constant for geometry data

Minimal complete definition

hasM, hasZ, geoType

Methods

hasM :: a -> Bool

hasZ :: a -> Bool

geoType :: a -> Word32

class Typeable a => Geometry a where

Type class which defined base DB operations

Minimal complete definition

putGeometry, getGeometry

Methods

putGeometry :: Putter a

getGeometry :: Get a

Serialization types

type Putter a = a -> Put

type Getter = ReaderT Header Get

data Header

Header record

Spatial Reference Identifier

type SRID = Maybe Int

Default reader and writer for EWKB data

writeEWKB :: Geometry a => a -> ByteString

readEWKB :: Geometry a => ByteString -> a

Function which can be used in user specific getters and putters

makeHeader :: GeoData a => SRID -> a -> Header

putHeader :: GeoData a => SRID -> a -> Put

putMaybe :: Maybe a -> Putter a -> Put

putPoint :: Putter (Double, Double, Maybe Double, Maybe Double)

putChainLen :: Putter Int

getHeader :: Get Header

getHeaderPre :: Get Header

lookGeoType :: Header -> Word32

mkGeom :: Header -> (SRID -> a -> b) -> Getter a -> Get b

skipHeader :: Getter ()

getPoint :: Getter (Double, Double, Maybe Double, Maybe Double)

getChainLen :: Getter Int

PostGIS constants

pgisPoint :: Word32

pgisLinestring :: Word32

pgisPolygon :: Word32

pgisMultiPoint :: Word32

pgisMultiLinestring :: Word32

pgisMultiPolygon :: Word32